+++ lib/entity/inventory.ts
@@ -47,8 +47,9 @@ export class EntityInventory extends Container implements I.HasCustomInspection 
     }
 
     /// @internal
-    public [I.customInspectSymbol](inspect: (value: any, opts?: I.InspectOptions) => PP.Doc,
-                                   stylise: (token: PP.Doc, type: I.TokenType) => PP.Doc): PP.Doc {
+    public override [I.customInspectSymbol](inspect: (value: any, opts?: I.InspectOptions) => PP.Doc,
+                                            stylise: (token: PP.Doc, type: I.TokenType) => PP.Doc
+                                           ): PP.Doc {
         const obj: any = {
             containerType: this.containerType,
             container: new Container(this.#raw.container!),